+++ assets/javascripts/kitten/components/cards/article-card.test.js
@@ -8,7 +8,7 @@ import { ButtonImage } from 'kitten/components/buttons/button-image'
 describe('<ArticleCard />', () => {
   let component
   let buttonImage
-  let cardImage
+  let title
 
   describe('by default', () => {
     beforeEach(() => {
@@ -36,8 +36,8 @@ describe('<ArticleCard />', () => {
       expect(component.find(Title)).toHaveLength(1)
     })
 
-    it('has 3 <Text /> components', () => {
-      expect(component.find(Text)).toHaveLength(3)
+    it('has 2 <Text /> components', () => {
+      expect(component.find(Text)).toHaveLength(2)
     })
 
     it('has <HorizontalStroke /> component', () => {
@@ -115,6 +115,29 @@ describe('<ArticleCard />', () => {
           cursor: 'crosshair',
         },
       })
+    })
+  })
+
+  describe('with title props', () => {
+    beforeEach(() => {
+      component = mount(
+        <ArticleCard
+          titleProps={{
+            tag: 'h2',
+            className: 'custom-class',
+          }}
+        />
+      )
+
+      title = component.find(Title).first()
+    })
+
+    it('has a title with h2 tag', () => {
+      expect(title).to.have.prop('tag', 'h2')
+    })
+
+    it('has a custom class on Title', () => {
+      expect(title).to.have.className('custom-class')
     })
   })
 })