+++ assets/javascripts/kitten/components/steppers/stepper-icon.js
@@ -6,9 +6,10 @@ import { CheckedIcon } from 'kitten/components/icons/checked-icon'
 const StepperIcon = ({ className, children }) => {
   const iconClassNames = classNames(
     'k-ButtonIcon',
-    'k-ButtonIcon--micro',
+    'k-ButtonIcon--nano',
     'k-ButtonIcon--hydrogen',
     'k-ButtonIcon--withoutHover',
+    'k-ButtonIcon--rounded',
     'k-Stepper__icon',
     className,
   )
@@ -22,7 +23,7 @@ const StepperIcon = ({ className, children }) => {
 
 export const StepperIconDefault = () => {
   return (
-    <StepperIcon className="k-ButtonIcon--empty--micro" />
+    <StepperIcon className="k-ButtonIcon--empty--nano" />
   )
 }
 
@@ -43,7 +44,7 @@ export const StepperIconInProgress = () => {
 export const StepperIconValidated = () => {
   const validatedClassNames = classNames(
     'k-Stepper__icon--validated',
-    'k-ButtonIcon--checked--micro',
+    'k-ButtonIcon--checked--nano',
   )
 
   return (